MVP local models
