+++ .streamlit/secrets.toml
@@ -1,0 +1,1 @@
+openai_api_key = "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"
