+++ alphastats/gui/utils/gpt_helper.py
@@ -1,16 +1,10 @@
 import copy
-from typing import Optional, Union, List
-from pathlib import Path
-import requests
+from typing import Union, List, Dict
 
-import time
-import random
 import json
 
 from Bio import Entrez
-from gprofiler import GProfiler
 
-import openai
 import pandas as pd
 import streamlit as st
 
@@ -19,62 +13,10 @@ from alphastats.plots.DimensionalityReduction import DimensionalityReduction
 
 Entrez.email = "lebedev_mikhail@outlook.com"  # Always provide your email address when using NCBI services.
 
-uniprot_fields = [
-    # Names & Taxonomy
-    "gene_names",
-    "organism_name",
-    "protein_name",
-    # Function
-    "cc_function",
-    "cc_catalytic_activity",
-    "cc_activity_regulation",
-    "cc_pathway",
-    "kinetics",
-    "ph_dependence",
-    "temp_dependence",
-    # Interaction
-    "cc_interaction",
-    "cc_subunit",
-    # Expression
-    "cc_tissue_specificity",
-    "cc_developmental_stage",
-    "cc_induction",
-    # Gene Ontology (GO)
-    "go",
-    "go_p",
-    "go_c",
-    "go_f",
-    # Pathology & Biotech
-    "cc_disease",
-    "cc_disruption_phenotype",
-    "cc_pharmaceutical",
-    "ft_mutagen",
-    "ft_act_site",
-    # Structure
-    "cc_subcellular_location",
-    "organelle",
-    "absorption",
-    # Publications
-    "lit_pubmed_id",
-    # Family & Domains
-    "protein_families",
-    "cc_domain",
-    "ft_domain",
-    # Protein-Protein Interaction Databases
-    "xref_biogrid",
-    "xref_intact",
-    "xref_mint",
-    "xref_string",
-    # Chemistry Databases
-    "xref_drugbank",
-    "xref_chembl",
-    "reviewed",
-]
-
 
 def get_subgroups_for_each_group(
     metadata: pd.DataFrame,
-) -> dict:
+) -> Dict:
     """
     Get the unique values for each column in the metadata file.
 
@@ -101,7 +43,7 @@ def get_unique_values_from_column(column: str, metadata: pd.DataFrame) -> List[s
         metadata (pd.DataFrame, optional): The metadata dataframe (which sample has which disease/treatment/condition/etc).
 
     Returns:
-        list[str]: A list of unique values from the column.
+        List[str]: A list of unique values from the column.
     """
     unique_values = metadata[column].unique().tolist()
     return [str(i) for i in unique_values]
@@ -135,23 +77,7 @@ def display_proteins(overexpressed: List[str], underexpressed: List[str]) -> Non
     st.markdown(full_html, unsafe_allow_html=True)
 
 
-def get_assistant_functions(
-    gene_to_prot_id_dict: dict,
-    metadata: pd.DataFrame,
-    subgroups_for_each_group: dict,
-) -> List[dict]:
-    """
-    Get a list of assistant functions for function calling in the ChatGPT model.
-    You can call this function with no arguments, arguments are given for clarity on what changes the behavior of the function.
-    For more information on how to format functions for Assistants, see https://platform.openai.com/docs/assistants/tools/function-calling
-
-    Args:
-        gene_to_prot_id_dict (dict, optional): A dictionary with gene names as keys and protein IDs as values.
-        metadata (pd.DataFrame, optional): The metadata dataframe (which sample has which disease/treatment/condition/etc).
-        subgroups_for_each_group (dict, optional): A dictionary with the column names as keys and a list of unique values as values. Defaults to get_subgroups_for_each_group().
-    Returns:
-        list[dict]: A list of assistant functions.
-    """
+def get_general_assistant_functions() -> List[Dict]:
     return [
         {
             "type": "function",
@@ -173,7 +99,55 @@ def get_assistant_functions(
         {
             "type": "function",
             "function": {
-                "name": "st.session_state.dataset.plot_intensity",
+                "name": "get_enrichment_data",
+                "description": "Get enrichment data for a list of differentially expressed genes",
+                "parameters": {
+                    "type": "object",
+                    "properties": {
+                        "difexpressed": {
+                            "type": "array",
+                            "items": {"type": "string"},
+                            "description": "A list of differentially expressed gene names to search for",
+                        },
+                        "organism_id": {
+                            "type": "string",
+                            "description": "The Uniprot organism ID to search in, e.g. 9606 for human",
+                        },
+                        "tool": {
+                            "type": "string",
+                            "description": "The tool to use for enrichment analysis",
+                            "enum": ["gprofiler", "string"],
+                        },
+                    },
+                    "required": ["difexpressed", "organism_id"],
+                },
+            },
+        },
+    ]
+
+
+def get_assistant_functions(
+    gene_to_prot_id_dict: Dict,
+    metadata: pd.DataFrame,
+    subgroups_for_each_group: Dict,
+) -> List[Dict]:
+    """
+    Get a list of assistant functions for function calling in the ChatGPT model.
+    You can call this function with no arguments, arguments are given for clarity on what changes the behavior of the function.
+    For more information on how to format functions for Assistants, see https://platform.openai.com/docs/assistants/tools/function-calling
+
+    Args:
+        gene_to_prot_id_dict (dict, optional): A dictionary with gene names as keys and protein IDs as values.
+        metadata (pd.DataFrame, optional): The metadata dataframe (which sample has which disease/treatment/condition/etc).
+        subgroups_for_each_group (dict, optional): A dictionary with the column names as keys and a list of unique values as values. Defaults to get_subgroups_for_each_group().
+    Returns:
+        list[dict]: A list of assistant functions.
+    """
+    return [
+        {
+            "type": "function",
+            "function": {
+                "name": "plot_intensity",
                 "description": "Create an intensity plot based on protein data and analytical methods.",
                 "parameters": {
                     "type": "object",
@@ -241,7 +215,7 @@ def get_assistant_functions(
         {
             "type": "function",
             "function": {
-                "name": "st.session_state.dataset.plot_sampledistribution",
+                "name": "plot_sampledistribution",
                 "description": "Generates a histogram plot for each sample in the dataset matrix.",
                 "parameters": {
                     "type": "object",
@@ -264,7 +238,7 @@ def get_assistant_functions(
         {
             "type": "function",
             "function": {
-                "name": "st.session_state.dataset.plot_volcano",
+                "name": "plot_volcano",
                 "description": "Generates a volcano plot based on two subgroups of the same group",
                 "parameters": {
                     "type": "object",
@@ -322,34 +296,7 @@ def get_assistant_functions(
                 },
             },
         },
-        {
-            "type": "function",
-            "function": {
-                "name": "get_enrichment_data",
-                "description": "Get enrichment data for a list of differentially expressed genes",
-                "parameters": {
-                    "type": "object",
-                    "properties": {
-                        "difexpressed": {
-                            "type": "array",
-                            "items": {"type": "string"},
-                            "description": "A list of differentially expressed gene names to search for",
-                        },
-                        "organism_id": {
-                            "type": "string",
-                            "description": "The Uniprot organism ID to search in",
-                        },
-                        "tool": {
-                            "type": "string",
-                            "description": "The tool to use for enrichment analysis",
-                            "enum": ["gprofiler", "string"],
-                        },
-                    },
-                    "required": ["difexpressed", "organism_id"],
-                },
-            },
-        },
-        {"type": "code_interpreter"},
+        # {"type": "code_interpreter"},
     ]
 
 
@@ -360,347 +307,7 @@ def perform_dimensionality_reduction(group, method, circle, **kwargs):
     return dr.plot
 
 
-def get_uniprot_data(
-    gene_name: str,
-    organism_id: str,
-    fields: List[str] = uniprot_fields,
-) -> dict:
-    """
-    Get data from UniProt for a given gene name and organism ID.
-
-    Args:
-        gene_name (str): The gene name to search for.
-        organism_id (str, optional): The organism ID to search in. Defaults to streamlit session state.
-        fields (list[str], optional): The fields to retrieve from UniProt. Defaults to uniprot_fields defined above.
-
-    Returns:
-        dict: The data retrieved from UniProt.
-    """
-    base_url = "https://rest.uniprot.org/uniprotkb/search"
-    query = f"(gene:{gene_name}) AND (reviewed:true) AND (organism_id:{organism_id})"
-
-    response = requests.get(
-        base_url, params={"query": query, "format": "json", "fields": ",".join(fields)}
-    )
-
-    if response.status_code != 200:
-        print(
-            f"Failed to retrieve data for {gene_name}. Status code: {response.status_code}"
-        )
-        print(response.text)
-        return None
-
-    data = response.json()
-
-    if not data.get("results"):
-        print(f"No UniProt entry found for {gene_name}")
-        return None
-
-    # Return the first result as a dictionary (assuming it's the most relevant)
-    data = data["results"][0]
-    for key, value in data.items():
-        print(f"data - {key}: {value}, {type(value)}")
-    return data
-
-
-def extract_data(data: dict) -> dict:
-    """
-    Extract relevant data from a UniProt entry.
-
-    Args:
-        data (dict): The data retrieved from UniProt.
-
-    Returns:
-        dict: The extracted data.
-    """
-    extracted = {}
-
-    # 1. Entry Type
-    extracted["entryType"] = data.get("entryType", None)
-
-    # 2. Primary Accession
-    extracted["primaryAccession"] = data.get("primaryAccession", None)
-
-    # 3. Organism Details
-    organism = data.get("organism", {})
-    extracted["organism"] = {
-        "scientificName": organism.get("scientificName", None),
-        "commonName": organism.get("commonName", None),
-        "taxonId": organism.get("taxonId", None),
-        "lineage": organism.get("lineage", []),
-    }
-
-    # 4. Protein Details
-    protein_description = data.get("proteinDescription", {})
-    recommended_name = (
-        protein_description.get("recommendedName", {})
-        .get("fullName", {})
-        .get("value", None)
-    )
-    alternative_names = [
-        alt_name["fullName"]["value"]
-        for alt_name in protein_description.get("alternativeNames", [])
-    ]
-    extracted["protein"] = {
-        "recommendedName": recommended_name,
-        "alternativeNames": alternative_names,
-        "flag": protein_description.get("flag", None),
-    }
-
-    # 5. Gene Details
-    genes = data.get("genes", [{}])[0]
-    extracted["genes"] = {
-        "geneName": genes.get("geneName", {}).get("value", None),
-        "synonyms": [syn["value"] for syn in genes.get("synonyms", [])],
-    }
-
-    # 6. Functional Comments
-    function_comments = [
-        text["value"]
-        for comment in data.get("comments", [])
-        if comment["commentType"] == "FUNCTION"
-        for text in comment.get("texts", [])
-    ]
-    extracted["functionComments"] = function_comments
-
-    # 7. Subunit Details
-    subunit_comments = [
-        text["value"]
-        for comment in data.get("comments", [])
-        if comment["commentType"] == "SUBUNIT"
-        for text in comment.get("texts", [])
-    ]
-    extracted["subunitComments"] = subunit_comments
-
-    # 8. Protein Interactions
-    interactions = []
-
-    for c in data.get("comments", []):
-        if c["commentType"] == "INTERACTION":
-            for interaction in c.get("interactions", []):
-                interactantOne = interaction["interactantOne"].get(
-                    "uniProtKBAccession", None
-                )
-                interactantTwo = interaction["interactantTwo"].get(
-                    "uniProtKBAccession", None
-                )
-
-                # Only append if both interactants are present
-                if interactantOne and interactantTwo:
-                    interactions.append(
-                        {
-                            "interactantOne": interactantOne,
-                            "interactantTwo": interactantTwo,
-                            "numberOfExperiments": interaction["numberOfExperiments"],
-                        }
-                    )
-    extracted["interactions"] = interactions
-
-    # 9. Subcellular Locations
-    subcellular_locations_comments = [
-        c["subcellularLocations"]
-        for c in data.get("comments", [])
-        if c["commentType"] == "SUBCELLULAR LOCATION"
-    ]
-    locations = [
-        location["location"]["value"]
-        for locations_comment in subcellular_locations_comments
-        for location in locations_comment
-    ]
-    extracted["subcellularLocations"] = locations
-
-    tissue_specificities = [
-        text["value"]
-        for comment in data.get("comments", [])
-        if comment["commentType"] == "TISSUE SPECIFICITY"
-        for text in comment.get("texts", [])
-    ]
-    extracted["tissueSpecificity"] = tissue_specificities
-
-    # 11. Protein Features
-    features = [
-        {
-            "type": feature["type"],
-            "description": feature["description"],
-            "location_start": feature["location"]["start"]["value"],
-            "location_end": feature["location"]["end"]["value"],
-        }
-        for feature in data.get("features", [])
-    ]
-    extracted["features"] = features
-
-    # 12. References
-    references = [
-        {
-            "authors": ref["citation"].get("authors", []),
-            "title": ref["citation"].get("title", ""),
-            "journal": ref["citation"].get("journal", ""),
-            "publicationDate": ref["citation"].get("publicationDate", ""),
-            "comments": [c["value"] for c in ref.get("referenceComments", [])],
-        }
-        for ref in data.get("references", [])
-    ]
-    extracted["references"] = references
-
-    # 13. Cross References
-    cross_references = [
-        {
-            "database": ref["database"],
-            "id": ref["id"],
-            "properties": {
-                prop["key"]: prop["value"] for prop in ref.get("properties", [])
-            },
-        }
-        for ref in data.get("uniProtKBCrossReferences", [])
-    ]
-    extracted["crossReferences"] = cross_references
-    return extracted
-
-
-def get_info(genes_list: List[str], organism_id: str) -> List[str]:
-    """
-    Get info from UniProt for a list of genes.
-
-    Args:
-        genes_list (list[str]): A list of gene names to search for.
-        organism_id (str, optional): The Uniprot organism ID to search in.
-
-    Returns:
-        list[str]: A list of gene functions."""
-    results = {}
-
-    for gene in genes_list:
-        result = get_uniprot_data(gene, organism_id)
-
-        # If result is retrieved for the gene, extract data and continue with the next gene
-        if result:
-            results[gene] = extract_data(result)
-            continue
-
-        # If no result is retrieved for the gene and the gene string does not contain a ";", continue with the next gene
-        if ";" not in gene:
-            print(f"Failed to retrieve data for {gene}")
-            continue
-
-        # If no result is retrieved for the gene and the gene string contains a ";", try to get data for each split gene
-        split_genes = gene.split(";")
-        for split_gene in split_genes:
-            result = get_uniprot_data(split_gene.strip(), organism_id)
-            if result:
-                print(
-                    f"Successfully retrieved data for {split_gene} (from split gene: {gene})"
-                )
-                results[gene] = extract_data(result)
-                break
-
-        # If still no result after trying split genes
-        if not result:
-            print(f"Failed to retrieve data for all parts of split gene: {gene}")
-            # TODO: Handle this case further if necessary
-
-    gene_functions = []
-    for gene in results:
-        if results[gene]["functionComments"]:
-            gene_functions.append(f"{gene}: {results[gene]['functionComments']}")
-        else:
-            gene_functions.append(f"{gene}: ?")
-
-    return gene_functions
-
-
-def get_functional_annotation_STRING(identifier, species_id="9606") -> pd.DataFrame:
-    """
-    Get functional annotation from STRING for a gene identifier.
-
-    Args:
-        identifier (str): A gene identifier.
-        species_id (str, optional): The Uniprot organism ID to search in.
-
-    Returns:
-        pd.DataFrame: The functional annotation data.
-    """
-    url = f"https://string-db.org/api/json/enrichment?identifiers={identifier}&species={int(species_id)}&caller_identity=alphapeptstats"
-    response = requests.get(url)
-
-    if response.status_code == 200:
-        data = response.json()
-        df = pd.DataFrame(data)
-        return df
-    else:
-        print(f"Request failed with status code {response.status_code}")
-        return None
-
-
-def get_functional_annotation_GProfiler(identifiers: List[str]) -> pd.DataFrame:
-    """
-    Get functional annotation from g:Profiler for a list of gene identifiers.
-
-    Args:
-        identifiers (list[str]): A list of gene identifiers.
-
-    Returns:
-        pd.DataFrame: The functional annotation data.
-    """
-    gp = GProfiler(
-        user_agent="AlphaPeptStats",
-        return_dataframe=True,
-    )
-    df = gp.profile(query=identifiers)
-    return df
-
-
-def get_enrichment_data(
-    difexpressed: List[str], organism_id: str = 9606, tool: str = "gprofiler"
-) -> pd.DataFrame:
-    """
-    Get enrichment data for a list of differentially expressed genes.
-
-    Args:
-        difexpressed (list[str]): A list of differentially expressed genes.
-        organism_id (str, optional): The Uniprot organism ID to search in.
-        tool (str, optional): The tool to use for enrichment analysis.
-
-    Returns:
-        pd.DataFrame: The enrichment data.
-    """
-    enrichment_data = {}
-    assert tool in [
-        "gprofiler",
-        "string",
-    ], "Tool must be either 'gprofiler' or 'string'"
-    if tool == "gprofiler":
-        enrichment_data = get_functional_annotation_GProfiler(difexpressed)
-    else:
-        enrichment_data = get_functional_annotation_STRING(
-            "%0d".join(difexpressed), organism_id
-        )
-
-    return enrichment_data
-
-
-def get_gene_function(gene_name: Union[str, dict], organism_id: str = None) -> str:
-    """
-    Get the gene function and description by UniProt lookup of gene identifier / name.
-
-    Args:
-        gene_name (Union[str, dict]): Gene identifier / name for UniProt lookup.
-        organism_id (str): The UniProt organism ID to search in.
-
-    Returns:
-        str: The gene function and description.
-    """
-    if not organism_id:
-        organism_id = st.session_state["organism"]
-    if type(gene_name) == dict:
-        gene_name = gene_name["gene_name"]
-    result = get_uniprot_data(gene_name, organism_id)
-    if result and extract_data(result)["functionComments"]:
-        return str(extract_data(result)["functionComments"])
-    else:
-        return "No data found"
-
-
-def turn_args_to_float(json_string: Union[str, bytes, bytearray]) -> dict:
+def turn_args_to_float(json_string: Union[str, bytes, bytearray]) -> Dict:
     """
     Turn all values in a JSON string to floats if possible.
 
@@ -740,160 +347,3 @@ def get_gene_to_prot_id_mapping(gene_id: str) -> str:
         if gene_id in gene.split(";"):
             return prot_id
     return gene_id
-
-
-def wait_for_run_completion(
-    client: openai.OpenAI, thread_id: int, run_id: int, check_interval: int = 2
-) -> Optional[list]:
-    """
-    Wait for a run and function calls to complete and return the plots, if they were created by function calling.
-
-    Args:
-        client (openai.OpenAI): The OpenAI client.
-        thread_id (int): The thread ID.
-        run_id (int): The run ID.
-        check_interval (int, optional): The interval to check for run completion. Defaults to 2 seconds.
-
-    Returns:
-        Optional[list]: A list of plots, if any.
-    """
-    artefacts = []
-    while True:
-        run_status = client.beta.threads.runs.retrieve(
-            thread_id=thread_id, run_id=run_id
-        )
-        assistant_functions = {
-            "create_intensity_plot",
-            "perform_dimensionality_reduction",
-            "create_sample_histogram",
-            "st.session_state.dataset.plot_volcano",
-            "st.session_state.dataset.plot_sampledistribution",
-            "st.session_state.dataset.plot_intensity",
-            "st.session_state.dataset.plot_pca",
-            "st.session_state.dataset.plot_umap",
-            "st.session_state.dataset.plot_tsne",
-            "get_enrichment_data",
-        }
-        if run_status.status == "completed":
-            print("Run is completed!")
-            if artefacts:
-                print("Returning artefacts")
-                return artefacts
-            break
-        elif run_status.status == "requires_action":
-            print("requires_action", run_status)
-            print(
-                [
-                    st.session_state.plotting_options[i]["function"].__name__
-                    for i in st.session_state.plotting_options
-                ]
-            )
-            tool_calls = run_status.required_action.submit_tool_outputs.tool_calls
-            tool_outputs = []
-            for tool_call in tool_calls:
-                print("### calling:", tool_call.function.name)
-                if tool_call.function.name == "get_gene_function":
-                    print(
-                        type(tool_call.function.arguments), tool_call.function.arguments
-                    )
-                    prompt = json.loads(tool_call.function.arguments)["gene_name"]
-                    gene_function = get_gene_function(prompt)
-                    tool_outputs.append(
-                        {
-                            "tool_call_id": tool_call.id,
-                            "output": gene_function,
-                        },
-                    )
-                elif (
-                    tool_call.function.name
-                    in [
-                        st.session_state.plotting_options[i]["function"].__name__
-                        for i in st.session_state.plotting_options
-                    ]
-                    or tool_call.function.name in assistant_functions
-                ):
-                    args = tool_call.function.arguments
-                    args = turn_args_to_float(args)
-                    print(f"{tool_call.function.name}(**{args})")
-                    artefact = eval(f"{tool_call.function.name}(**{args})")
-                    artefact_json = artefact.to_json()
-
-                    tool_outputs.append(
-                        {"tool_call_id": tool_call.id, "output": artefact_json},
-                    )
-                    artefacts.append(artefact)
-
-            if tool_outputs:
-                _run = client.beta.threads.runs.submit_tool_outputs(
-                    thread_id=thread_id, run_id=run_id, tool_outputs=tool_outputs
-                )
-        else:
-            print("Run is not yet completed. Waiting...", run_status.status, run_id)
-            time.sleep(check_interval)
-
-
-def send_message_save_thread(client: openai.OpenAI, message: str) -> Optional[list]:
-    """
-    Send a message to the OpenAI ChatGPT model and save the thread in the session state, return plots if GPT called a function to create them.
-
-    Args:
-        client (openai.OpenAI): The OpenAI client.
-        message (str): The message to send to the ChatGPT model.
-
-    Returns:
-        Optional[list]: A list of plots, if any.
-    """
-    message = client.beta.threads.messages.create(
-        thread_id=st.session_state["thread_id"], role="user", content=message
-    )
-
-    run = client.beta.threads.runs.create(
-        thread_id=st.session_state["thread_id"],
-        assistant_id=st.session_state["assistant"].id,
-    )
-    try:
-        plots = wait_for_run_completion(client, st.session_state["thread_id"], run.id)
-    except KeyError as e:
-        print(e)
-        plots = None
-    messages = client.beta.threads.messages.list(
-        thread_id=st.session_state["thread_id"]
-    )
-    st.session_state.messages = []
-    for num, message in enumerate(messages.data[::-1]):
-        role = message.role
-        if message.content:
-            content = message.content[0].text.value
-        else:
-            content = "Sorry, I was unable to process this message. Try again or change your request."
-        st.session_state.messages.append({"role": role, "content": content})
-    if not plots:
-        return
-    return plots
-
-
-def try_to_set_api_key(api_key: str = None) -> None:
-    """
-    Checks if the OpenAI API key is available in the environment / system variables.
-    If the API key is not available, saves the key to secrets.toml in the repository root directory.
-
-    Args:
-        api_key (str, optional): The OpenAI API key. Defaults to None.
-
-    Returns:
-        None
-    """
-    if api_key and "api_key" not in st.session_state:
-        st.session_state["openai_api_key"] = api_key
-        secret_path = Path(st.secrets._file_paths[-1])
-        secret_path.parent.mkdir(parents=True, exist_ok=True)
-        with open(secret_path, "w") as f:
-            f.write(f'openai_api_key = "{api_key}"')
-        openai.OpenAI.api_key = api_key
-        return
-    try:
-        openai.OpenAI.api_key = st.secrets["openai_api_key"]
-    except:
-        st.write(
-            "OpenAI API key not found in environment variables. Please enter your API key to continue."
-        )
